@@ -20,10 +20,10 @@
         "bin/zf-dependency-injection-cache-warmup"
     ],
     "require": {
-        "php": "~7.4.0 || ~8.0.0 || ~8.1.0 || ~8.2.0",
+        "php": "~8.1.0 || ~8.2.0",
         "laminas/laminas-servicemanager": "^2.7 | ^3.0",
         "laminas/laminas-modulemanager": "^2.7",
-        "psr/simple-cache": "^1.0"
+        "psr/simple-cache": "^1.0 | ^2.0 | ^3.0"
     },
     "autoload": {
         "psr-4": {
