@@ -14,12 +14,9 @@ use Psr\Container\ContainerInterface;
  */
 final class Inject implements AnnotationInterface
 {
-    /**
-     * @var string
-     */
-    public $value;
+    public string $value;
 
-    public function __invoke(ContainerInterface $container)
+    public function __invoke(ContainerInterface $container): mixed
     {
         return $container->get($this->value);
     }
