@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Reinfi\DependencyInjection\Annotation;
 
+use Doctrine\ORM\EntityRepository;
 use Psr\Container\ContainerInterface;
 use Reinfi\DependencyInjection\Exception\AutoWiringNotPossibleException;
 
@@ -34,7 +35,7 @@ final class InjectDoctrineRepository extends AbstractAnnotation
         $this->entity = $values['value'];
     }
 
-    public function __invoke(ContainerInterface $container)
+    public function __invoke(ContainerInterface $container): EntityRepository
     {
         $container = $this->determineContainer($container);
 
