@@ -21,7 +21,7 @@ final class Inject implements InjectionInterface
         $this->value = $value;
     }
 
-    public function __invoke(ContainerInterface $container)
+    public function __invoke(ContainerInterface $container): mixed
     {
         return $container->get($this->value);
     }
